+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All future changes from this point on will be documented in this file.
 - wfuzz from [@xmendez](https://github.com/xmendez/wfuzz)
 - LinEnum from [@cheetz](https://github.com/cheetz/LinEnum)
 - MeterSSH from [@trustedsec](https://github.com/trustedsec/meterssh)
+- The-Backdoor-Factory from [@secretsquirrel](https://github.com/secretsquirrel/the-backdoor-factory)
+- SigThief from [@secretSquirrel](https://github.com/secretsquirrel/SigThief)
+- Veil-Framework from [@Veil-Framework](https://github.com/Veil-Framework/Veil)
 
 ### Changed
 - Functionalized the entire script so that every git installation is in its own function. I did this to help the user specify which packages won't be installed by simply commenting out a function call at the bottom of the script.